Tidy ChatClient by reorganising query grouped channels
