+++ CHANGELOG.md
@@ -4,9 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 # Upcoming
 
 ### ✅ Added
-- Add `ChatChannelListController.prefill(channels:completion:)` for priming controller-local channel data before the first synchronize call while preserving normal pagination, observation, and offline refresh behavior
-- Add `ChatClient.groupedQueryChannels(limit:watch:presence:)` to fetch grouped channel groups as `GroupedChannels`, preserving backend group keys and exposing normalized per-group channels and unread counts for integrators
-- Add optional `groupedUnreadChannels` data to grouped unread websocket events and persist it on `CurrentChatUser` for integrators
+- Add `ChatChannelListController.prefill(channels:completion:)` for priming controller-local channel data
+- Add `ChatClient.queryGroupedChannels(limit:watch:presence:)` to fetch grouped channels with per group unread counts
+- Add optional `groupedUnreadChannels` data to relevant web-socket events and to `CurrentChatUser`
 
 ### 🔄 Changed
 - Make grouped channels decoding tolerate missing `unread_count` and `unread_channels` fields in group buckets, matching the current OpenAPI schema
